# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CRM API - Leads & Customers
  description: API สำหรับ Agent OpenCLaw - จัดการ Leads และ Customers
  version: 1.0.0

servers:
  - url: http://localhost:3000/crm
    description: Local
  - url: https://your-crm-domain.com/crm
    description: Production

security:
  - sessionCookie: []

paths:
  /api/entities/check:
    get:
      summary: ตรวจสอบว่า externalId มีในระบบแล้วหรือยัง
      description: เรียกก่อนเพิ่ม Lead/Customer เพื่อป้องกันข้อมูลซ้ำ
      operationId: checkEntityExists
      parameters:
        - name: externalId
          in: query
          required: true
          schema:
            type: string
          description: รหัสลูกค้าภายนอก (Customer ID)
        - name: brandId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  externalId:
                    type: string
                  brandId:
                    type: string
                  inLeads:
                    type: boolean
                  inCustomers:
                    type: boolean
                  lead:
                    type: object
                    nullable: true
                  customer:
                    type: object
                    nullable: true

  /api/leads:
    get:
      summary: รายการ Leads
      parameters:
        - name: brandId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: stage
          in: query
          schema:
            type: string
            enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: OK
    post:
      summary: สร้าง Lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                brandId:
                  type: string
                title:
                  type: string
                externalId:
                  type: string
                  nullable: true
                firstName:
                  type: string
                  nullable: true
                lastName:
                  type: string
                  nullable: true
                firstNameTh:
                  type: string
                  nullable: true
                lastNameTh:
                  type: string
                  nullable: true
                titlePrefixTh:
                  type: string
                  nullable: true
                phone:
                  type: string
                  nullable: true
                email:
                  type: string
                  nullable: true
                source:
                  type: string
                  nullable: true
                stage:
                  type: string
                  enum: [prospecting, qualification, proposal, negotiation, closed_won, closed_lost]
                interest:
                  type: string
                  nullable: true
                notes:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Created

  /api/leads/{id}:
    get:
      summary: ดู Lead รายละเอียด
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Not found
    put:
      summary: แก้ไข Lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
    delete:
      summary: ลบ Lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

  /api/leads/{id}/convert:
    post:
      summary: แปลง Lead เป็น Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

  /api/leads/{id}/chat-logs:
    get:
      summary: ดึง Chat logs ของ Lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    post:
      summary: เพิ่ม Chat log
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, senderName, message]
              properties:
                channel:
                  type: string
                  enum: [LINE, Facebook, Instagram, WhatsApp, WeChat, Telegram, Phone, Email, Walk-in, Other]
                senderName:
                  type: string
                message:
                  type: string
                sentAt:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Created

  /api/customers:
    get:
      summary: รายการ Customers
      parameters:
        - name: brandId
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, prospect]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: OK
    post:
      summary: สร้าง Customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                brandId:
                  type: string
                name:
                  type: string
                externalId:
                  type: string
                  nullable: true
                firstName:
                  type: string
                  nullable: true
                lastName:
                  type: string
                  nullable: true
                firstNameTh:
                  type: string
                  nullable: true
                lastNameTh:
                  type: string
                  nullable: true
                phone:
                  type: string
                  nullable: true
                email:
                  type: string
                  nullable: true
                source:
                  type: string
                  nullable: true
                stage:
                  type: string
                  nullable: true
                status:
                  type: string
                  enum: [active, inactive, prospect]
      responses:
        '201':
          description: Created

  /api/customers/{id}:
    get:
      summary: ดู Customer รายละเอียด
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
    put:
      summary: แก้ไข Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
    delete:
      summary: ลบ Customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK

  /api/customers/{id}/chat-logs:
    get:
      summary: ดึง Chat logs ของ Customer
    post:
      summary: เพิ่ม Chat log

  /api/sources:
    get:
      summary: รายการ Sources
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: OK